ContainerProperties: Refactors Vector Embedding and Indexing Policy Interfaces to Mark Them as Public for GA (#4845)
# Pull Request Template

## Description

The purpose of this PR is to mark the new `VectorEmbeddingPolicy` in the
`ContainerProperties` as a public surface interface for `GA` release,
and introducing new `VectorIndexes` in the `IndexingPolicy` to enable
Vector Similarity Search in Cosmos DB ecosystem.

Relevant PRs for the vector similarity work: 

- [ContainerProperties: Adds Vector Embedding and Indexing
Policy](https://github.com/Azure/azure-cosmos-dotnet-v3/pull/4379)
- [ContainerProperties: Refactors Vector Embedding and Indexing Policy
Interfaces to Mark Them as Public for
Preview](https://github.com/Azure/azure-cosmos-dotnet-v3/pull/4486)
- [VectorIndexDefinition: Adds Support for Partitioned
DiskANN](https://github.com/Azure/azure-cosmos-dotnet-v3/pull/4792)

## Type of change

Please delete options that are not relevant.

- [x] New feature (non-breaking change which adds functionality)

## Closing issues

To automatically close an issue: closes #4825