[Internal] Query: Removes unneeded check on supported query features (#4868)
This PR

-Upgrades to the 3.37.1 version of the direct package
-Adds additional testing on ORDER BY RANK RRF with VectorDistance in
HybridSearchQueryTests SanityTests
-Removes an unneeded (and problematic) check for unsupported features,
which is not needed by the SDK. Previously, this code assumed there
would always be a queryInfo object, however with hybrid search, this may
not always be the case. But in addition, this check was not needed, so
it is removed.

Please delete options that are not relevant.

- [] Bug fix (non-breaking change which fixes an issue)

Co-Authored-By: Kiran Kumar Kolli <kirankk@microsoft.com>