3.45.1: Adds new SDK versions and contract files (#4870)
This bumps the SDK GA version from `3.45.0` to `3.45.1` and preview
version from `3.46.0-preview.0` to `3.46.0-preview.1`. No changes in the
public GA contract. The preview contract changes are given below:

![image](https://github.com/user-attachments/assets/009825a1-9765-44ef-94bf-a65e951d75a9)

This version also bumps up the `Cosmos.Direct` package to `3.37.1`,
which includes the below changes:

- Currently Gateway is running into a `NullReferenceException` when
hybrid query is supplied for execution. This is because for hybrid
queries, `QueryInfo` in the plan is null (and instead
`HybridSearchQueryInfo` is non-null). This is resulting in `500
(InternalServerError`) for JS SDK scenarios. The fix adds a `null` check
for this.

Please delete options that are not relevant.

- [x] New feature (non-breaking change which adds functionality)

To automatically close an issue: closes #IssueNumber