Query: Fixes (workaround) for query plan issue where placeholder index does not start at zero (#4885)
## Description

During query plan generation, we were using the loop index index over
order by expressions as the index for placeholders within component
queries. However, VectorDistance expressions do not require
placeholders. As a result the output placeholder indexes may not always
start at 0. The fix for this is currently in PR on the gateway.

We are putting this workaround in the SDK until the gateway fix is
deployed.

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)