Hedging: Adds support for writes on multi region accounts (#4706)
# Pull Request Template

## Description

Adds support to use hedging on write requests for multimaster accounts.
Note that this does come with the caveat that there will be more 409
errors thrown by the SDK. This is expected and applications that adapt
this feature should be prepared to handle these exceptions.

This feature will also be an opt in feature for Availability Strategies
and users will need to especially specify that in the availability
strategy definition.

```
ItemRequestOptions requestOptions = new ItemRequestOptions
{
    AvailabilityStrategy = new CrossRegionHedgingAvailabilityStrategy(
        threshold: TimeSpan.FromMilliseconds(100),
        thresholdStep: TimeSpan.FromMilliseconds(50),
        multiMasterWritesEnabled: true)
};
```

## Type of change

Please delete options that are not relevant.

- [] New feature (non-breaking change which adds functionality)

## Closing issues

To automatically close an issue: closes #4457

---------

Co-authored-by: Fabian Meiswinkel <fabian@meiswinkel.com>
Co-authored-by: Kevin Pilch <kevinpi@microsoft.com>