[Internal] Pipelines: Adds FI pipeline YMLs (#4909)
# Pull Request Template

## Description

Adds YML files needed for fault injection release.

These files might need to be revisited in a separate PR, as there is no
way to test the new pipeline without these files being in master first.

## Type of change

Please delete options that are not relevant.

- [] New feature (non-breaking change which adds functionality)

## Closing issues

To automatically close an issue: closes #IssueNumber