[Internal] Pipelines: Fixes FaultInjection Release Pipelines (#4913)
# Pull Request Template

## Description

Fixes FaultInjection Release Pipelines:

- Removes build-fault-injection-test.yml, this file is unneeded and
including it just runs the same tests twice increasing cogs
- Updates FaultInjection Tests, these tests were originally made using a
session consistency account, the live account we use for integration
testing in running with strong consistency. Changes to the tests needed
to be made in order for them to run on strong consistency.

![image](https://github.com/user-attachments/assets/e12b056a-a558-47e3-bda3-2c9a234743a5)


## Type of change

Please delete options that are not relevant.

- [] Bug fix (non-breaking change which fixes an issue)

## Closing issues

To automatically close an issue: closes #IssueNumber