[Internal] Query: Adds Query Advisor SDK capabilities (#4714)
# Pull Request Template

## Description

This PR adds the capability to request and receive Query Advice in
QueryRequests. The response from BE is an URL-encoded JSON string that
will get decoded to UTF8 text.

To request query advice, use the new
`QueryRequestOpton.PopulateQueryAdvice`
Once this is set to true, see the advice via `FeedResponse.QueryAdvice`
property

Note: the test is marked Ignore because the emulator used in pipeline
does not support Query Advisor yet. Furthermore, this request option is
internal and not yet public to wait for the deployment

## Type of change

Please delete options that are not relevant.

- [] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [] This change requires a documentation update

## Closing issues

To automatically close an issue: closes #IssueNumber

---------

Co-authored-by: leminh98 <leminh@microsoft.com>