Query : Fixes OFFSET, LIMIT, TOP data types to match those coming from query plan (#4939)
## Query : Fixes OFFSET, LIMIT, TOP data types to match those coming
from query plan

This change ensures that the data types of OFFSET, LIMIT and TOP fields
on QueryInfo and HybridSearchQueryInfo objects match those coming from
the query plan. Currently the queryplan/service uses ULONG which has a
range of 0x00000000 to 0xFFFFFFFF. Since QueryInfo fields use int?, this
causes an overflow for a value that's larger than 0x7FFFFFFF for any of
these fields.

## Type of change

Please delete options that are not relevant.

- [x] Bug fix (non-breaking change which fixes an issue)
- [] New feature (non-breaking change which adds functionality)
- [] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [] This change requires a documentation update