Per Partition Automatic Failover: Adds requirement to enable application region or application preferred regions for PPAF (#4987)
# Pull Request Template

## Description

ClientOptions now validates that either ApplicationRegion or
ApplicationPreferredRegion is mandatory when PartitionLevelFailover is
enabled.

Live testing was also done independently to validate this change and
that PPAF works for gateway interactions, with PPAF, SDK will now
failover to secondary regions for gateway interactions.

## Type of change

Please delete options that are not relevant.

- [] New feature (non-breaking change which adds functionality)

## Closing issues

To automatically close an issue: closes #4974