[Internal] 3.47.0: Fixes Release branch (#5007)
[Internal] Open Telemetry : Fixes preview pipeline test for open
telemetry (#5005)

## Description

As part of this PR, fixing test pipeline.

**Root Cause:** Metrics generation tests were designed to validate only
metrics functionality with dimensions. However, with distributed tracing
enabled by default in the preview package, the test began asserting
traces generated by the distributed tracing feature. **Resolution:**
Disabled the distributed tracing feature in tests to ensure
compatibility.

## Type of change

- [] Bug fix (non-breaking change which fixes an issue)

# Pull Request Template

## Description

Please include a summary of the change and which issue is fixed. Include
samples if adding new API, and include relevant motivation and context.
List any dependencies that are required for this change.

## Type of change

Please delete options that are not relevant.

- [] Bug fix (non-breaking change which fixes an issue)
- [] New feature (non-breaking change which adds functionality)
- [] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [] This change requires a documentation update

## Closing issues

To automatically close an issue: closes #IssueNumber

Co-authored-by: Sourabh Jain <sourabhjain@microsoft.com>