Direct code: Adds a direct package that contains a fix for Faster detection of broken transport connections. (#5020)
# Pull Request Template
Direct package upgrade to 3.37.10
Also contains an emulator test to verify connection management during
aggressive timeouts.
Brings in the fix from Direct package:


https://msdata.visualstudio.com/CosmosDB/_git/CosmosDB/pullrequest/1574317?_a=files

## Description

Please include a summary of the change and which issue is fixed. Include
samples if adding new API, and include relevant motivation and context.
List any dependencies that are required for this change.

## Type of change

Please delete options that are not relevant.

- [] Bug fix (non-breaking change which fixes an issue)
- [] New feature (non-breaking change which adds functionality)
- [] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [] This change requires a documentation update

## Closing issues

To automatically close an issue: closes #IssueNumber