3.47.1 hotfix: Adds build props for sdk-hotfix-release-3.47.1 (#5026)
# Pull Request Template
Brings in the build.props changes and contracts from 3.47.1 Hotfix
release.

- Build.props change from the commit
`74e91c43c9b70d0936a7098ee02f06e6c561d674`

## Description

NO CONTRACT CHANGES



![image](https://github.com/user-attachments/assets/570c4b8f-39e0-40d7-bdb5-5a76a1e4bedd)


## Type of change

Please delete options that are not relevant.

- [] Bug fix (non-breaking change which fixes an issue)
- [] New feature (non-breaking change which adds functionality)
- [] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [] This change requires a documentation update

## Closing issues

To automatically close an issue: closes #IssueNumber

# Pull Request Template

## Description

Please include a summary of the change and which issue is fixed. Include
samples if adding new API, and include relevant motivation and context.
List any dependencies that are required for this change.

## Type of change

Please delete options that are not relevant.

- [] Bug fix (non-breaking change which fixes an issue)
- [] New feature (non-breaking change which adds functionality)
- [] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [] This change requires a documentation update

## Closing issues

To automatically close an issue: closes #IssueNumber