[Internal] Pipelines: Adds fix to rolling pipelines (#5044)
# Pull Request Template

## Description

Currently the rolling pipelines are failing because of a forgotten
environment variable. This PR will add that variable and also add a
preview run to this pipeline. Ultimately, the goal here will be to catch
any flaky tests that exist in the SDK to ensure faster PR turn around
time, and a smoother release process for the SDK.

## Type of change

Please delete options that are not relevant.

- [] Bug fix (non-breaking change which fixes an issue)


## Closing issues

To automatically close an issue: closes #IssueNumber