[Internal]: Fixes NullRef in TransportHandler when client is concurrently disposed. (#5062)
# Pull Request Template

## Description

Fixes a race conditions where StoreProxy returned can be null when the
client is getting disposed at the same time.

NOTE: Testing this is not easy due to this ebing a race condition. I
evaluated using the ExceptionLess tests (which is using
MockTransportHandler) - but it was not sutainable - because the race
condition to hit the NullRef is tiny. Given that throwing
ObjectDisposedException happens in other places regularly in this code
path I think it is acceptable to not add a unit tests trying to provoke
this edge condition.

## Type of change

Please delete options that are not relevant.

- [x] Bug fix (non-breaking change which fixes an issue)
- [] New feature (non-breaking change which adds functionality)
- [] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [] This change requires a documentation update

## Closing issues

To automatically close an issue: closes #IssueNumber