[Internal]Thin Client Integration: Adds Classes and Methods for Transport Serialization of RNTBD Payload (#5019)
# Pull Request Template

## Description

This pull request introduces the ProxyStoreClient, ThinClientStoreModel,
and ThinClientTransportSerializer classes as part of SDK-Thinclient
integration. These additions enhance the SDK's functionality by enabling
operations through a ThinClient proxy, which includes applying session
tokens, resolving partition key ranges, and delegating requests to the
ProxyStoreClient. The ThinClientTransportSerializer provides
serialization and deserialization of requests and responses to and from
the RNTBD protocol for the ThinClient scenario.
Also, included functional unit tests for these classes.

- **ThinClientStoreModel:** This class implements the
`ProcessMessageAsync()` method from the `IStoreModel` interface, to send
the request message via the proxy store client. This store model
abstracts out the implementation of the thin client interaction using
the new Http 2 protocol.

- **ThinClientStoreClient:** This client is responsible for interacting
with the thin client over Http2, using the thin client uri.

- **ThinClientTransportSerializer:** This class provides static methods
for serializing and deserializing requests and responses to and from the
RNTBD protocol for the ThinClient scenario.

## Type of change

Please delete options that are not relevant.

- [X] New feature (non-breaking change which adds functionality)


## Closing issues

To automatically close an issue: closes #4571

---------

Co-authored-by: Debdatta Kunda <87335885+kundadebdatta@users.noreply.github.com>