[Internal] PPAF: Adds Partition Level Failover for Gateway Mode (#5087)
# Pull Request Template

## Description

This PR introduces the partition level failover (PPAF + PPCB) for
Gateway Mode. The changes are mainly done in the `GatewayStoreModel`
which leverages the `GlobalPartitionEndpointManager` to apply the
partition level override.

Since the gateway mode data plane requests are retried through the
`ClientRetryPolicy`, the inbuild logic for the `Direct` mode suffices
the need for triggering the partition level fail over on certain error
codes.

As a part of the changes, we also introduced a new timeout policy
`HttpTimeoutPolicyForPartitionFailover` to fail fast the gateway
requests.

## Type of change

Please delete options that are not relevant.

- [x] New feature (non-breaking change which adds functionality)

## Closing issues

To automatically close an issue: closes #4982