Diagnostics: Fixes bug where some overloaded substatus codes are displayed incorrectly in diagnostics. (#5106)
# Pull Request Template

## Description

This pull request includes several changes to the
`Microsoft.Azure.Cosmos` project, focusing on improving the handling and
testing of sub-status codes in the tracing functionality. The most
significant changes include the addition of a new method to map
sub-status codes to more descriptive strings, updates to unit tests to
validate these changes.

### Improvements to sub-status code handling:

*
[`Microsoft.Azure.Cosmos/src/Tracing/TraceWriter.TraceJsonWriter.cs`](diffhunk://#diff-2d305bb945ccb69fdb9eb3404e3d1016302f0bc22a5d32c879513834ed9e9c0fL397-R397):
Added the `GetSubStatusCodeString` method to map sub-status codes to
more descriptive strings based on the status code. Updated the
`Visit(StoreResult storeResult)` method to use this new mapping method.
[[1]](diffhunk://#diff-2d305bb945ccb69fdb9eb3404e3d1016302f0bc22a5d32c879513834ed9e9c0fL397-R397)
[[2]](diffhunk://#diff-2d305bb945ccb69fdb9eb3404e3d1016302f0bc22a5d32c879513834ed9e9c0fR455-R564)

## Type of change

Please delete options that are not relevant.

- [] Bug fix (non-breaking change which fixes an issue)

## Closing issues

To automatically close an issue: closes #5104